refactor: move extractDailyFractions to shared dailyAttribution module
Move the canonical extractDailyFractions() implementation from cli/src/helpers.ts
to a new shared module vscode-extension/src/dailyAttribution.ts so all editors
(Copilot CLI, VS Code, VS Code JSONL, and any future adapter) use the same
attribution logic from a single source of truth.

- New: vscode-extension/src/dailyAttribution.ts
  Exports extractDailyFractions(content, isJsonl, fallbackDate): Record<string,number>
  Handles all three formats: Copilot CLI JSONL, VS Code delta JSONL, VS Code JSON

- cli/src/helpers.ts: import extractDailyFractions from dailyAttribution (remove local copy)

- vscode-extension/src/ecosystemAdapter.ts: add optional getDailyFractions?() method
  to IEcosystemAdapter interface so new adapters can expose per-request timestamps
  for accurate multi-day attribution without needing ecosystem-specific branches in
  processSessionFile. Falls back to mtime when not implemented.
